+++ package.json
@@ -18,11 +18,12 @@
     "react-dom": "^18.1.0",
     "react-helmet-async": "^1.3.0",
     "react-icons": "^4.4.0",
+    "react-intersection-observer": "^9.4.0",
     "react-query": "^3.39.1",
     "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.1",
-    "react-spring-bottom-sheet": "^3.4.1",
     "react-slick": "^0.29.0",
+    "react-spring-bottom-sheet": "^3.4.1",
     "recoil": "^0.7.4",
     "slick-carousel": "^1.8.1",
     "styled-components": "^5.3.5",
