+++ src/components/EventList/index.tsx
@@ -1,20 +1,13 @@
-import React from "react";
+import React, { useEffect } from "react";
 import { useInfiniteQuery } from "react-query";
+import { useInView } from "react-intersection-observer";
 import { StyledList } from "../../styles/eventListStyle";
 import EventListItem from "./EventListItem";
-import { supabase } from "../../supabaseClient";
-
-const ITEMS_PER_PAGE = 20;
-
-const fetchEvents = async ({ pageParam = 1 }) => {
-	const endAt = pageParam * ITEMS_PER_PAGE;
-	const startAt = endAt - ITEMS_PER_PAGE === 0 ? endAt - ITEMS_PER_PAGE : endAt - ITEMS_PER_PAGE + 1;
-
-	const { data } = await supabase.from("events").select("*").range(startAt, endAt);
-	return data;
-};
+import { fetchEvents } from "../../apis";
+import { ITEMS_PER_PAGE } from "../../shared/constants";
 
 const EventList = () => {
+	const { ref, inView } = useInView();
 	const { data, hasNextPage, fetchNextPage } = useInfiniteQuery("events", fetchEvents, {
 		getNextPageParam: (lastPage, pages) => {
 			if (lastPage && lastPage?.length < ITEMS_PER_PAGE) {
@@ -24,13 +17,22 @@ const EventList = () => {
 		},
 	});
 
-	// todo: flat()으로 리팩토링
+	useEffect(() => {
+		if (inView && hasNextPage) {
+			fetchNextPage();
+		}
+	}, [inView, hasNextPage, fetchNextPage]);
+
+	// todo: flat()으로 리팩토링 또는 데이터 구조 변경
 	return (
 		<StyledList>
-			{data?.pages.map((page) => page?.map((item) => <EventListItem event={item} key={item.id} />))}
-			<button type="button" disabled={!hasNextPage} onClick={() => fetchNextPage()}>
-				더보기
-			</button>
+			{data?.pages.map((page) =>
+				page?.map((item) => (
+					<div key={item.id} ref={ref}>
+						<EventListItem event={item} key={item.id} />
+					</div>
+				))
+			)}
 		</StyledList>
 	);
 };
