+++ src/shared/constants.ts
@@ -1,0 +1,3 @@
+export const ITEMS_PER_PAGE = 20;
+
+export default {};
